refactor: move api logic to src/api dir
