+++ src/README.md
@@ -1,5 +1,10 @@
 ## ディレクトリ設計
 
+### api
+
+pages/api/ 配下に作成した API の詳細処理を配置する。
+pages/api/ 配下のエンドポイントと一対一でディレクトリを構成すること。
+
 ### assets
 
 静的なアセットのうち、ビルドプロセス内に組み込んだほうがいいものを配置する。
@@ -21,7 +26,8 @@ TBD
 
 ### templates
 
-TBD
+各ルーティング配下の UI を配置する。
+pages/ 配下のルーティングと一対一でディレクトリを構成すること。
 
 ### types
